finalizing

+++ parse.c
@@ -22,7 +22,7 @@ int json_p(struct response *resp, struct clock_data *c_data){
     carbon_deadline_1_timestamp_s = (char*)json_object_get_string(carbon_deadline_1_timestamp);
     renewables_1_rate_i = json_object_get_double(renewables_1_rate);
     renewables_1_timestamp_s = (char*)json_object_get_string(renewables_1_timestamp);
-    renewables_1_initial = (long double)json_object_get_double(renewables_1_initial_j);
+    renewables_1_initial = (double)json_object_get_double(renewables_1_initial_j);
 
     c_data->carbon_deadline_1 = (char*)malloc(*carbon_deadline_1_timestamp_s);
     c_data->renewables_1_timestamp = (char*)malloc(*renewables_1_timestamp_s);
